HarmonyView/harmonyview/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pyodbc==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
websockets==12.0
python-multipart==0.0.6
